Move ui to console package
